Updated Rental class to utilize the Sql class

+++ vs_solution/Assessment2/Form_Rental.xaml.cs
@@ -9,7 +9,7 @@ namespace Assessment2
     public partial class Form_Rental : Window
     {
         private ulong vehicleId;
-        private int rentalId;
+        private ulong rentalId;
         private decimal dEndOdometer;
         private DateTime? dtEndDate = null;
         /// <summary>
